requirements.txt
Django==5.2.1
djangorestframework==3.16.0
drf-yasg==1.21.10
torch==2.7.0
torchaudio==2.7.0
transformers==4.52.3
pydub==0.25.1
librosa==0.11.0
numpy==2.1.3
soundfile==0.13.1
mutagen==1.47.0